feat: add API for creating temporary users
