+++ noweekend-core/core-api/src/main/kotlin/noweekend/core/api/controller/HealthController.kt
@@ -1,9 +1,6 @@
 package noweekend.core.api.controller
 
 import io.swagger.v3.oas.annotations.Hidden
-import noweekend.core.domain.auth.TestUserService
-import noweekend.core.domain.auth.UserWithToken
-import noweekend.core.support.response.ApiResponse
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
@@ -11,19 +8,10 @@ import org.springframework.web.bind.annotation.RestController
 
 @Hidden
 @RestController
-class HealthController(
-    private val testUserService: TestUserService,
-) {
+class HealthController {
     @GetMapping("/health")
     fun health(): ResponseEntity<*> {
         return ResponseEntity.status(HttpStatus.OK)
             .body("ok")
-    }
-
-    @GetMapping("/test-gen")
-    fun createTestGen(): ApiResponse<UserWithToken> {
-        return ApiResponse.success(
-            testUserService.testUserGen(),
-        )
     }
 }
